fix: move the types DeviceStatus, StatusCode to timeline-state-resolver-types
This is both because I think they serve a better purpose there, but also so that they are exported by timeline-state-resolver (they where not before)
